inha win8 metrowin contribution / "porting mode" programming - many IDEs in different platforms and OSes... Eu te amo Thommy, Thomas Borregaard Sørensen!!
Mummi contribution!!!! Feliz 61 anos!!

at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 7 Ultimate 64-bit Service Pack 1 - with Portuguese - Brazil - Language for non-Unicode programs - ~10Mpbs NET Virtua Internet at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

｛
｝

                  -   -   -
Mummi contribution!!, eu te amo demais!!

-----8nnnnnnnn8
++++++++++

+++ os/metrowin_dib.cpp
@@ -2631,7 +2631,7 @@ namespace metrowin
 
       }
 
-      int iArea = area();
+      int64_t iArea = area();
 
       byte * p = (byte *) m_pcolorref;
